further cleanup and city stage
