+++ src/main/java/com/canoo/cog/ui/welcome/WelcomeController.java
@@ -23,15 +23,18 @@ package com.canoo.cog.ui.welcome;
 
 import java.util.List;
 
+import com.canoo.cog.solver.CityNode;
+import com.canoo.cog.solver.LittleBetterSolverEver;
+import com.canoo.cog.solver.Solver;
+import com.canoo.cog.solver.SonarToStrategyConerter;
 import com.canoo.cog.sonar.SonarException;
 import com.canoo.cog.sonar.SonarService;
 import com.canoo.cog.sonar.model.CityModel;
 import com.canoo.cog.sonar.model.SonarProject;
-import com.canoo.cog.ui.citybuilder.CityBuilder;
+import com.canoo.cog.ui.city.CityStage;
 import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.fxml.FXML;
-import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.PasswordField;
 import javafx.scene.control.TableColumn;
@@ -39,7 +42,6 @@ import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.input.KeyCode;
-import javafx.stage.Stage;
 
 class WelcomeController {
 
@@ -117,15 +119,20 @@ class WelcomeController {
 
     private void loadCodeCity() {
         try {
+            // Get City Data from Sonar
             SonarProject selectedItem = (SonarProject) projectTable.getSelectionModel().getSelectedItem();
             CityModel cityData = sonarService.getCityData(selectedItem.getKey());
-            CityBuilder cityBuilder = new CityBuilder(cityData);
-            Scene scene = cityBuilder.build();
-            Stage stage = new Stage();
-            stage.setTitle(cityBuilder.getTitle());
-            stage.setScene(scene);
-            stage.setFullScreen(true);
-            stage.show();
+
+            // Solve problem with Solver
+            final int STREET_SIZE = 6;
+            Solver solver = new LittleBetterSolverEver();
+            CityNode resultNode = new SonarToStrategyConerter().convertCityToNode(cityData);
+            solver.solveProblem(resultNode, STREET_SIZE);
+
+            // Start City Application
+            final CityStage cityStage = new CityStage(cityData, resultNode);
+            cityStage.startStage();
+
         } catch (SonarException e) {
             // show error message in Sonar
         }
