+++ src/main/resources/com/canoo/cog/ui/city/model/style/td.xcf
